fixed cout rounding error

+++ src/parser/parser.cpp
@@ -251,8 +251,8 @@ int parse_schedule(Schedule &schedule, std::string &filename) {
         std::istringstream start_date("1/Jun/2027 00:00:00.000");
         timepoint tp_start;
         start_date >> date::parse("%d/%b/%Y %T", tp_start);
-        int start_int = std::stoi(start_str);
-        int end_int = std::stoi(end_str);
+        long int start_int = std::stol(start_str);
+        long int end_int = std::stol(end_str);
         const timepoint start = tp_start + std::chrono::milliseconds(start_int);
         const timepoint end = tp_start + std::chrono::milliseconds(end_int);
